+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Dump CSRs using `regs` subcommand
+- Show firmware size info when flashing
 
 ### Fixed